requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
iniconfig==2.0.0
isodate==0.6.1
lxml==4.9.2
mpv==1.0.3
packaging==23.1
pluggy==1.0.0
pycountry==22.3.5
pycryptodome==3.17
PySocks==1.7.1
pytest==7.3.1
python-dotenv==1.0.0
requests==2.28.2
six==1.16.0
streamlink==5.4.0
urllib3==1.26.15
websocket-client==1.5.1
